script/requirements_Jetson_Nano.txt
absl-py==0.12.0
actionlib==1.12.1
Adafruit-Blinka==5.13.1
adafruit-circuitpython-busdevice==5.0.4
adafruit-circuitpython-pca9685==3.3.3
adafruit-circuitpython-register==1.9.3
Adafruit-PlatformDetect==2.28.0
Adafruit-PureIO==1.1.8
angles==1.9.12
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.4.2
astunparse==1.6.3
beautifulsoup4==4.6.0
blinker==1.4
bondpy==1.8.5
Brlapi==0.6.6
cachetools==4.2.1
camera-calibration-parsers==1.11.13
catkin==0.7.29
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
cupshelpers==1.0
cv-bridge==1.13.0
Cython==0.29.22
decorator==4.1.2
defer==1.0.6
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
dynamic-reconfigure==1.6.3
feedparser==5.2.1
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
futures==3.1.1
gast==0.3.3
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
graphsurgeon==0.4.5
grpcio==1.32.0
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
image-geometry==1.13.0
importlib-metadata==3.7.2
interactive-markers==1.11.5
isort==5.7.0
itsdangerous==1.1.0
Jetson.GPIO==2.0.15
Jinja2==2.11.3
joint-state-publisher==1.12.15
kdl-parser-py==1.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
message-filters==1.14.10
mock==3.0.5
numpy==1.19.4
oauth==1.0.1
oauthlib==3.1.0
onboard==1.4.1
opt-einsum==3.3.0
packaging==20.9
PAM==0.4.2
pandas==0.22.0
pbr==5.5.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyftdi==0.52.9
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyserial==3.5
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
python-qt-binding==0.4.3
pytz==2018.3
pyusb==1.1.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
resource-retriever==1.12.6
rosbag==1.14.10
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.10
roslaunch==1.14.10
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.10
rosmake==1.14.9
rosmaster==1.14.10
rosmsg==1.14.10
rosnode==1.14.10
rosparam==1.14.10
rospy==1.14.10
rosservice==1.14.10
rostest==1.14.10
rostopic==1.14.10
rosunit==1.14.9
roswtf==1.14.10
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rsa==4.7.2
rviz==1.13.16
scikit-build==0.11.1
scipy==0.19.1
SecretStorage==2.3.1
sensor-msgs==1.12.8
simplejson==3.13.2
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0+nv21.2
tensorflow-estimator==2.4.0
tensorrt==7.1.3.0
termcolor==1.1.0
testresources==2.0.1
tf==1.12.1
tf-conversions==1.12.1
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
toml==0.10.2
topic-tools==1.14.10
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
uff==0.6.9
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urdfdom-py==0.4.5
urllib3==1.22
urwid==2.0.1
wadllib==1.3.2
webencodings==0.5
Werkzeug==1.0.1
wrapt==1.12.1
xacro==1.13.9
xkit==0.0.0
youtube-dl==2018.3.14
zipp==3.4.1
zope.interface==4.3.2
